pyproject.toml
[project]
name = "ngs360-server"
version = "0.1.0"
description = "NGS360 API Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "authlib>=1.6.11",
    "bcrypt>=4.0.0,<6.0.0",
    "boto3>=1.40.30",
    "cryptography>=45.0.6",
    "email-validator>=2.1.0",
    "fastapi[standard]>=0.116.1",
    "gunicorn>=21.0.0",
    "httpx>=0.27.0",
    "mako>=1.3.10",
    "opensearch-py>=3.0.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.27",
    "pytz>=2025.2",
    "sample-sheet>=0.13.0",
    "smart-open>=7.3.1",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "flake8",
    "pytest",
    "pytest-cov"
]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
    "mock>=5.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]
